requirements.txt
aiogram==3.1.1
aiogram-i18n==1.4
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-core==2.6.3